Move action resources to dedicated pkg
